mm: vmscan: move shrinker_debugfs_remove() before synchronize_srcu()
The debugfs_remove_recursive() will wait for debugfs_file_put() to return,
so there is no need to put it after synchronize_srcu() to wait for the rcu
read-side critical section to exit.

Just move it before synchronize_srcu(), which is also convenient to put
the heavy synchronize_srcu() in the delayed work later.

Link: https://lkml.kernel.org/r/20230601132714.2540707-1-qi.zheng@linux.dev
Signed-off-by: Qi Zheng <zhengqi.arch@bytedance.com>
Cc: Roman Gushchin <roman.gushchin@linux.dev>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>